% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_cs.R
\name{reset_cs}
\alias{reset_cs}
\title{Reset Selenium WebDriver service}
\usage{
reset_cs()
}
\description{
Reset Selenium WebDriver service
}
\examples{
auto_cs(dest_dir = "C:/Users/Dell/Desktop/test")
reset_cs()
}
