% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdauto.R
\name{wdauto}
\alias{wdauto}
\title{Title}
\usage{
wdauto(dest_dir)
}
\arguments{
\item{dest_dir}{The path of chrome browser dirver and selenium server}
}
\description{
Description
}
\examples{
wdauto(dest_dir = "C:/Users/Dell/Desktop/test/test")
}
