% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cs.R
\name{auto_cs}
\alias{auto_cs}
\title{Start Selenium WebDriver service}
\usage{
auto_cs(dest_dir)
}
\arguments{
\item{dest_dir}{The path of chrome browser driver and selenium server}
}
\description{
Start Selenium WebDriver service
}
\examples{
auto_cs(dest_dir = "C:/Users/Dell/Desktop/test")
}
