% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cd.R
\name{get_cd}
\alias{get_cd}
\title{Title}
\usage{
get_cd(dest_dir)
}
\arguments{
\item{dest_dir}{The path of chrome browser dirver and selenium server}
}
\description{
Description
}
\examples{
get_cd(dest_dir = "C:/Users/Dell/Desktop/test/test")
}
